refactor : ChessGame의 board를 ChessBoard 클래스로 분리
